Move `overlaps` test suite to `boundary`
